requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
gunicorn==23.0.0
blinker==1.9.0
click==8.1.8
Flask==2.2.5
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.0.1
Werkzeug==2.2.3
WTForms==3.2.1
